Title: 정수 사이의 배수, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 251023/정수 사이의 배수/multiplication-between-numbers.cpp
@@ -19,7 +19,7 @@ int main()
         }
     }
 
-    fs_avg = fs_sum / fs_cnts;
+    fs_avg = (double)fs_sum / fs_cnts;
     cout << fixed;
     cout << fs_sum << ' ' << setprecision(1) << fs_avg;
     return 0;
